Refactor: Add mesage type in most places
